refactor: index.js 코드 리팩토링
- index.js 내용을 UI.js 파일로 이동
